requirements.txt
# Core dependencies
llvmlite>=0.40.0
pyyaml>=6.0.1
typing-extensions>=4.5.0

# Web interface
flask>=2.3.3
flask-cors>=4.0.0
python-dotenv>=1.0.0

# AI features
openai>=1.3.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1

# Documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0 